# internal/docs/swagger.yaml
basePath: /v1
definitions:
  auth.SignInRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  auth.SignInResponse:
    properties:
      access_token:
        type: string
      id_token:
        type: string
    type: object
  auth.SignUpRequest:
    properties:
      department_id:
        type: string
      email:
        type: string
      facility_id:
        type: string
      password:
        type: string
      phone_number:
        type: string
      position_id:
        type: string
      team_id:
        type: string
      username:
        type: string
    required:
    - department_id
    - facility_id
    - password
    - position_id
    - team_id
    - username
    type: object
  auth.SignUpResponse:
    properties:
      department:
        type: string
      email:
        type: string
      facility:
        type: string
      id:
        type: string
      name:
        type: string
      phone_number:
        type: string
      policies:
        items:
          $ref: '#/definitions/policy.Policy'
        type: array
      position:
        type: string
      team:
        type: string
    type: object
  common.ErrorResponse:
    properties:
      code:
        type: string
      description:
        type: string
      msg:
        type: string
    type: object
  department.DepartmentResponse:
    properties:
      created_at:
        type: string
      facility_id:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  policy.CreatePolicyRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  policy.CreatePolicyResponse:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  policy.Policy:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  policy.PolicyResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  position.CreatePositionRequest:
    properties:
      name:
        type: string
      policy_ids:
        items:
          type: string
        type: array
    required:
    - name
    - policy_ids
    type: object
  position.PolicyDto:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  position.PositionResponse:
    properties:
      created_at:
        type: string
      facility_id:
        type: string
      id:
        type: string
      name:
        type: string
      policies:
        items:
          $ref: '#/definitions/position.PolicyDto'
        type: array
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: BuddyのAPIサーバー
  title: Buddy-API
  version: "1.0"
paths:
  /auth/signin:
    post:
      consumes:
      - application/json
      parameters:
      - description: Sign In Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.SignInRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/auth.SignInResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: サインイン用エンドポイント
      tags:
      - Auth
  /auth/signup:
    post:
      consumes:
      - application/json
      parameters:
      - description: Sign Up Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/auth.SignUpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: サインアップ用エンドポイント
      tags:
      - Auth
  /departments/{department_id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Department ID
        in: path
        name: department_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/department.DepartmentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: 単一の部署を取得する
      tags:
      - Department
  /facilities/{facility_id}/departments:
    get:
      consumes:
      - application/json
      parameters:
      - description: Facility ID
        in: path
        name: facility_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/department.DepartmentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: 施設IDに紐づく部署を取得する
      tags:
      - Department
  /facilities/{facility_id}/positions:
    get:
      consumes:
      - application/json
      parameters:
      - description: Facility ID
        in: path
        name: facility_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/position.PositionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: 施設IDに紐づく役職を取得する
      tags:
      - Position
    post:
      consumes:
      - application/json
      parameters:
      - description: Create Position Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/position.CreatePositionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/position.PositionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: 施設に紐づく役職を作成する
      tags:
      - Position
...
  /policies:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/policy.PolicyResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: ポリシー一覧を取得する
      tags:
      - Policy
    post:
      consumes:
      - application/json
      parameters:
      - description: Create Policy Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/policy.CreatePolicyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/policy.CreatePolicyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: ポリシーを作成する
      tags:
      - Policy
  /policies/{policy_id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Policy ID
        in: path
        name: policy_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/policy.PolicyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: 単一のポリシーを取得する
      tags:
      - Policy
  /positions/{position_id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Position ID
        in: path
        name: position_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/position.PositionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: 単一の役職を取得する
      tags:
      - Position
...
swagger: "2.0"